chore: massive refactor
