+++ src/posts/2015-03-09-mobile-device-etiquette.md
