+++ src/posts/2015-10-06-immediate-functions.md
