+++ src/posts/2015-10-13-simple-variadic-behavior.md
