+++ src/posts/2018-01-23-ramda-chops-function-currying.md
