+++ src/posts/2018-01-26-ramda-chops-function-composition.md
