+++ src/posts/2018-01-31-ramda-chops-safely-accessing-properties.md
