+++ src/posts/2018-02-08-ramda-chops-map-filter-reduce.md
