+++ src/posts/2020-02-20-javascript-write-a-functional-programming-style-map-function.md
