+++ src/posts/2023-01-30-how-to-lose-functional-programming-at-work.md
