+++ src/posts/2023-03-20-one-does-not-simply-install-ghcup-on-mac-m1.md
