+++ src/templates/default_nz.html
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <title>$title$</title>
+
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width,initial-scale=1">
+
+    <meta name="description" content="$desc$">
+    $if(author)$<meta name="author" content="$author$">$endif$
+    $if(keywords)$<meta name="keywords" content="$keywords$">$endif$
+    <meta name="google-site-verification" content="QX4bn65yOXBfgnwxlO09yIBc1H8blur-Erx3lmsDVhU">
+    <meta http-equiv="origin-trial" content="Am4BZ0c7GMyB72dgo/Ny2FfIFscXhYMoN+CVe4jduWh24FvsaCwf7kjZzHzfrJXtIilyZVAEKRxOItGLY7lvkAgAAABSeyJvcmlnaW4iOiJodHRwczovL3JvYmVydHdwZWFyY2UuY29tOjQ0MyIsImZlYXR1cmUiOiJQb3J0YWxzIiwiZXhwaXJ5IjoxNjAzNTQ2NDc3fQ==">
+
+    <meta property="og:site_name" content="$siteName$">
+    <meta property="og:title" content="$title$">
+    <meta property="og:url" content="$root$$url$">
+    <meta property="og:description" content="$desc$">
+    $if(image)$<meta property="og:image" content="$root$$image$">$endif$
+    $if(type)$
+    <meta property="og:type" content="$type$">
+    $else$
+    <meta property="og:type" content="website">
+    $endif$
+
+    $if(image)$
+    <meta property="twitter:card" content="summary_large_image">
+    <meta property="twitter:image" content="$root$$image$">
+    $endif$
+    <meta property="twitter:site" content="$siteName$">
+    <meta property="twitter:title" content="$title$">
+    <meta property="twitter:description" content="$desc$">
+    $if(authorTwitter)$<meta property="twitter:creator" content="$authorTwitter$">$endif$
+
+    <link rel="icon" href="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 100 100'><text y='0.9em' font-size='90'>💾</text></svg>">
+    <link rel="canonical" href="$root$$url$">
+    <link rel="alternate" href="./atom.xml" title="$feedTitle$" type="application/atom+xml">
+    <link rel="alternate" href="./rss.xml" title="$feedTitle$" type="application/rss+xml">
+  </head>
+  <body>
+    $body$
+  </body>
+</html>
