+++ ssg/ssg.cabal
@@ -21,7 +21,14 @@ executable hakyll-site
                    , text >= 1 && < 3
                    , time >= 1.8
                    , time-locale-compat >= 0.1
-  other-modules:   Text.HTML.TagSoup.Compressor
+
+  other-modules:     Hakyll.Site.Configuration
+                   , Hakyll.Site.CustomFields
+                   , Hakyll.Site.Feed
+                   , Hakyll.Site.Post
+                   , Hakyll.Site.Rules
+                   , Hakyll.Site.Sitemap
+                   , Text.HTML.TagSoup.Compressor
   ghc-options:     -Wall
                    -Wcompat
                    -Widentities
